+++ src/m_config.c
@@ -2380,7 +2380,7 @@ default_t defaults[] = {
     "input_map_blink",
     NULL, NULL,
     {0}, {UL,UL}, input, ss_keys, wad_no,
-    "key to make automap markers blink",
+    "key to temporarily highlight points of interest on the automap",
     input_map_blink, { {INPUT_KEY, 'b'} }
   },